Added mongoDb connection and rabbitMQ in all services

+++ MicroservicesCalculator/CalculatorApi/BLL/ExpressionManager.cs
@@ -6,7 +6,7 @@ namespace CalculatorAPI.BLL;
 
 public class ExpressionManager(IPublishEndpoint publishEndpoint) : IExpressionManager
 {
-    public async Task<double> Calculate(string expression, CancellationToken cancellationToken)
+    public async Task<decimal> Calculate(string expression, CancellationToken cancellationToken)
     {
         Stack<string> expressionAsStackInfix = Parse(expression);
         var expressionAsStackPostfix = ToPostfixForm(expressionAsStackInfix);
@@ -75,7 +75,7 @@ public class ExpressionManager(IPublishEndpoint publishEndpoint) : IExpressionMa
 
         foreach (var stackValue in expression)
         {
-            if (double.TryParse(stackValue, out _))
+            if (decimal.TryParse(stackValue, out _))
             {
                 finalStack.Push(stackValue);
             }
@@ -132,12 +132,12 @@ public class ExpressionManager(IPublishEndpoint publishEndpoint) : IExpressionMa
         }
     }
 
-    private async Task<double> CalculateResult(Stack<string> expressionAsStackPostfix, CancellationToken cancellationToken)
+    private async Task<decimal> CalculateResult(Stack<string> expressionAsStackPostfix, CancellationToken cancellationToken)
     {
-        Stack<double> calculationStack = new Stack<double>();
+        Stack<decimal> calculationStack = new Stack<decimal>();
         foreach (var stackValue in expressionAsStackPostfix)
         {
-            if (double.TryParse(stackValue, out var operand))
+            if (decimal.TryParse(stackValue, out var operand))
             {
                 calculationStack.Push(operand);
             }
